+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Detect chunked `Transfer-Encoding` case-insensitively
 - Reject HTTP/1.1 requests without a valid `Host` header with 400
 
 ## [0.10.0] - 2026-07-07